refactor Program
